+++ axum-extra/CHANGELOG.md
@@ -19,13 +19,16 @@ and this project adheres to [Semantic Versioning].
 - **change:** axum-extra's MSRV is now 1.60 ([#1239])
 - **added:** Add Protocol Buffer extractor and response ([#1239])
 - **added:** Add `Either*` types for combining extractors and responses into a
-  single type
+  single type ([#1263])
+- **added:** Add sync constructors to `CookieJar`, `PrivateCookieJar`, and
+  `SignedCookieJar` so they're easier to use in custom middleware
 
 [#1086]: https://github.com/tokio-rs/axum/pull/1086
 [#1119]: https://github.com/tokio-rs/axum/pull/1119
 [#1170]: https://github.com/tokio-rs/axum/pull/1170
 [#1214]: https://github.com/tokio-rs/axum/pull/1214
 [#1239]: https://github.com/tokio-rs/axum/pull/1239
+[#1263]: https://github.com/tokio-rs/axum/pull/1263
 
 # 0.3.5 (27. June, 2022)